+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmtaskresource.class.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+require_once (strtr(realpath(dirname(dirname(__FILE__))), '\\', '/') . '/rmtaskresource.class.php');
+class rmTaskResource_mysql extends rmTaskResource {}
+?>