+++ core/components/romanescobackyard/model/romanescobackyard/rmtaskresource.class.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+class rmTaskResource extends rmTask {}
+?>